refactor to compute navbar ACLs using registry content
This is a bit invasive, as code dependencies have to be reorganized
to gain access to the registry during configuration of multiple catalog
types, in order to generate a list of appropriate group IDs for
ACLs which enable certain navbar menu items for CFDE-CC or DCC staff.

- Removes datapackage module CLI stub.
- Adds release module with CLI stub for release catalog prep/reconfig.
- Retains submission module CLI for review catalog reconfig.
- Retains registry module CLI for registry catalog reconfig.
